+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 1.0.0: Update iOS SDK, Stablize APIs
+* replace Tangram es with another ios sdk
+* add `startLocationUpdating`,`stopLocationUpdating` for external control of user location
+* fix some bugs
+* migrate to wasm for web
+* add
 ### 1.0.0-rc.6: update dependencies
 ### 1.0.0-rc.5: fix userlocation in android
 * fix user location tracking in android side (bug:#507)